# docs/swagger.yaml
basePath: /
definitions:
...
  schema_templates.CreateSchemaTemplateRequest:
    properties:
      created_by:
        type: string
      generation_type:
        type: string
      is_active:
        type: boolean
      locked_at:
        type: string
      schema_json:
        type: object
    type: object
  schema_templates.SchemaTemplate:
    description: Schema template with versioning and activation support
    properties:
      created_at:
        example: "2026-01-19T03:40:00Z"
        type: string
      created_by:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      generation_type:
        example: CLASSIFICATION
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_active:
        example: true
        type: boolean
      locked_at:
        example: "2026-01-19T03:40:00Z"
        type: string
      schema_json:
        type: object
      version:
        example: 1
        type: integer
    type: object
...
  subjects.SubSubject:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      subject_id:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
  subjects.Subject:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      sub_subjects:
        items:
          $ref: '#/definitions/subjects.SubSubject'
        type: array
      updated_at:
        type: string
      url:
        type: string
    type: object
  subjects.SubjectForSelection:
    properties:
      display_name:
        type: string
      full_name:
        type: string
      id:
        type: string
      parent_id:
        type: string
      url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API with password OAuth2 authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms/
  title: Learning API
  version: "1.0"
paths:
...
  /eval-items/{id}/reviews:
    get:
      consumes:
      - application/json
      description: Admin-only. List review history for an eval item.
      parameters:
      - description: Eval Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "501":
          description: not implemented
          schema:
            type: string
      security:
      - OAuth2:
        - read
      summary: List eval item reviews
      tags:
      - reviews
    post:
      consumes:
      - application/json
      description: Teacher+Admin. Submit a review verdict for an eval item.
      parameters:
      - description: Eval Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "501":
          description: not implemented
          schema:
            type: string
      security:
      - OAuth2:
        - write
      summary: Create eval item review
      tags:
      - reviews
...
  /schema-templates:
    get:
      description: Get all schema templates for a specific generation type
      parameters:
      - description: Generation type (CLASSIFICATION or QUESTIONS)
        in: query
        name: generation_type
        required: true
        type: string
      responses:
        "200":
          description: List of schema templates
          schema:
            items:
              $ref: '#/definitions/schema_templates.SchemaTemplate'
            type: array
        "400":
          description: Bad request - missing generation_type
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - read
      summary: List schema templates by generation type
      tags:
      - Schema Templates
    post:
      consumes:
      - application/json
      description: Create a new schema template (immutable - cannot edit existing)
      parameters:
      - description: Schema template request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema_templates.CreateSchemaTemplateRequest'
      responses:
        "201":
          description: Created schema template
          schema:
            $ref: '#/definitions/schema_templates.SchemaTemplate'
        "400":
          description: Bad request - invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - write
      summary: Create new schema template
      tags:
      - Schema Templates
  /schema-templates/{id}:
    get:
      description: Retrieve a specific schema template by its UUID
      parameters:
      - description: Template ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Schema template details
          schema:
            $ref: '#/definitions/schema_templates.SchemaTemplate'
        "400":
          description: Bad request - invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Template not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - read
      summary: Get schema template by ID
      tags:
      - Schema Templates
  /schema-templates/{id}/activate:
    post:
      description: Mark a schema template version as active (deactivates other versions)
      parameters:
      - description: Template ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Activated schema template
          schema:
            $ref: '#/definitions/schema_templates.SchemaTemplate'
        "400":
          description: Bad request - invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - write
      summary: Activate schema template
      tags:
      - Schema Templates
  /schema-templates/generation-type/{generationType}:
    get:
      description: Retrieve the currently active schema template for a specific generation
        type
      parameters:
      - description: Generation type (CLASSIFICATION or QUESTIONS)
        in: path
        name: generationType
        required: true
        type: string
      responses:
        "200":
          description: Active schema template
          schema:
            $ref: '#/definitions/schema_templates.SchemaTemplate'
        "400":
          description: Bad request - missing generation_type
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Active template not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - read
      summary: Get active schema template by generation type
      tags:
      - Schema Templates
...
  /subjects:
    get:
      consumes:
      - application/json
      description: Retrieve all academic subjects with their nested sub-subjects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/subjects.Subject'
            type: array
        "500":
          description: internal server error
          schema:
            type: string
      security:
      - OAuth2:
        - read
      summary: List all subjects
      tags:
      - subjects
  /subjects/for-selection:
    get:
      consumes:
      - application/json
      description: Retrieve subjects formatted for selection UI (tag cloud style)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/subjects.SubjectForSelection'
            type: array
        "500":
          description: internal server error
          schema:
            type: string
      security:
      - OAuth2:
        - read
      summary: List subjects for test creation selection
      tags:
      - subjects
...
securityDefinitions:
  OAuth2:
    flow: password
    tokenUrl: /oauth/token
    type: oauth2
swagger: "2.0"